setup.py niceities
I figured a useful package ought to be easy to install, so I've added a
setup.py. Along the way, I turned the README into standard rst and took
some of the examples from the docstring to make a nicer Github
splashpage. Also adds LICENSE.rst placeholder and a MANIFEST.in for
installation.

One thing left for you to do is:

[ ] Choose a license, add it to setup.py and LICENSE.rst